// frontend/src/api/generated/model/transportConfigOperation.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * keen-pbr API
 * REST API for the keen-pbr policy-based routing daemon.
 * OpenAPI spec version: 3.0.0
 */
import type { TransportConfigOperationOperation } from './transportConfigOperationOperation';
import type { TransportSpec } from './transportSpec';

export interface TransportConfigOperation {
  operation: TransportConfigOperationOperation;
  tag?: string;
  transport?: TransportSpec;
}